refactor: fixes lint
