+++ .eslintignore
@@ -1,11 +1,12 @@
 next-env.d.ts
 next.config.js
-node_modules
-manifests
-lang
-compiled-lang
-.waypoint
-.git
-.github
-.next
-dist
+
+node_modules/*
+manifests/*
+lang/*
+compiled-lang/*
+.waypoint/*
+.git/*
+.github/*
+.next/*
+dist/*
