--- .eslintrc.predev
@@ -1,9 +1,0 @@
-{
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "./tsconfig.server.json"
-  },
-  "extends": [
-    "@budproj/eslint-config-intl"
-  ]
-}
