+++ package.json
@@ -17,7 +17,9 @@
     "format": "prettier --write 'src/**/*.ts'",
     "lint": "eslint . --ext ts,tsx",
     "lint:fix": "npm run lint -- --fix",
-    "intl:generate:id": "echo '\u001b[36m➤ \u001b[90mGenerating react-intl message ids...\u001b[39m' && npm run lint -- --config .eslintrc.predev --fix",
+    "lint:timing": "TIMING=1 eslint . --ext ts,tsx",
+    "lint:messages": "eslint **/messages.ts",
+    "intl:generate:id": "echo '\u001b[36m➤ \u001b[90mGenerating react-intl message ids...\u001b[39m' && npm run lint:messages",
     "intl:refresh": "run-s intl:generate:id intl:extract",
     "intl:extract": "formatjs extract './src/**/*.ts(x)?' --out-file lang/pt-BR.json --id-interpolation-pattern '[sha512:contenthash:base64:6]'",
     "intl:compile:ptBR": "formatjs compile lang/pt-BR.json --ast --out-file compiled-lang/pt-BR.json",
@@ -41,11 +43,11 @@
   },
   "devDependencies": {
     "@babel/preset-react": "^7.12.7",
-    "@budproj/eslint-config-base-ts": "^0.0.12",
-    "@budproj/eslint-config-intl": "^0.0.12",
-    "@budproj/eslint-config-prettier-ts": "^0.0.12",
-    "@budproj/eslint-config-react": "^0.0.12",
-    "@budproj/prettier-config": "^0.0.12",
+    "@budproj/eslint-config-base-ts": "^1.0.0",
+    "@budproj/eslint-config-intl": "^1.0.0",
+    "@budproj/eslint-config-prettier-ts": "^1.0.0",
+    "@budproj/eslint-config-react": "^1.0.0",
+    "@budproj/prettier-config": "^1.0.0",
     "@formatjs/cli": "^2.13.5",
     "@miragejs/graphql": "^0.1.9",
     "@tsconfig/recommended": "^1.0.1",
@@ -102,6 +104,7 @@
     "@types/react-input-mask": "^3.0.0",
     "accepts": "^1.3.7",
     "chalk": "^4.1.0",
+    "cross-env": "^7.0.3",
     "crypto": "^1.0.1",
     "date-fns": "^2.16.1",
     "deepmerge": "^4.2.2",
