+++ src/pages/key-results/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import KeyResultsPage from 'src/components/Pages/KeyResults'
 
 const KeyResultsIndex = () => <KeyResultsPage />
