+++ src/state/recoil/key-result/icon/messages.ts
@@ -1,0 +1,126 @@
+import { defineMessages } from 'react-intl'
+
+import { KeyResultIconDesc } from './types'
+
+export const iconDescMessages: KeyResultIconDesc = defineMessages({
+  Activity: {
+    defaultMessage: 'Um ícone arredondado com um gráfico de linha ao centro',
+    id: '44uWtn',
+    description:
+      'The desc message for the Activity icon. This icon is used to ilustrate our key results',
+  },
+  Bookmark: {
+    defaultMessage: 'Um ícone de marcador de página',
+    id: '/vJzGe',
+    description:
+      'The desc message for the Bookmark icon. This icon is used to ilustrate our key results',
+  },
+  Calendar: {
+    defaultMessage: 'Um ícone de calendário',
+    id: 'lm9HAr',
+    description:
+      'The desc message for the Calendar icon. This icon is used to ilustrate our key results',
+  },
+  Delete: {
+    defaultMessage: 'Um ícone de lata de lixo',
+    id: 'NADPP1',
+    description:
+      'The desc message for the Delete icon. This icon is used to ilustrate our key results',
+  },
+  Discovery: {
+    defaultMessage: 'Um ícone de bussúla',
+    id: '03yPkn',
+    description:
+      'The desc message for the Discovery icon. This icon is used to ilustrate our key results',
+  },
+  Document: {
+    defaultMessage: 'Um ícone de documento de texto',
+    id: 'A5ct4X',
+    description:
+      'The desc message for the Document icon. This icon is used to ilustrate our key results',
+  },
+  EditSquare: {
+    defaultMessage: 'Um ícone de quadrado arredondado com um lápis',
+    id: 'ui441M',
+    description:
+      'The desc message for the EditSquare icon. This icon is used to ilustrate our key results',
+  },
+  Folder: {
+    defaultMessage: 'Um ícone de pasta',
+    id: '/VrdI0',
+    description:
+      'The desc message for the Folder icon. This icon is used to ilustrate our key results',
+  },
+  Game: {
+    defaultMessage: 'Um ícone de controle de video game',
+    id: 'P/ujgo',
+    description:
+      'The desc message for the Game icon. This icon is used to ilustrate our key results',
+  },
+  Graph: {
+    defaultMessage: 'Um ícone de gráfico de pizza',
+    id: 'ZXpNJ2',
+    description:
+      'The desc message for the Graph icon. This icon is used to ilustrate our key results',
+  },
+  Heart: {
+    defaultMessage: 'Um ícone de coração',
+    id: 'RWaMNF',
+    description:
+      'The desc message for the Heart icon. This icon is used to ilustrate our key results',
+  },
+  Location: {
+    defaultMessage: 'Um ícone de pin de localização',
+    id: 'XNkzYe',
+    description:
+      'The desc message for the Location icon. This icon is used to ilustrate our key results',
+  },
+  Message: {
+    defaultMessage: 'Um ícone de mensagem',
+    id: '1ss4IX',
+    description:
+      'The desc message for the Message icon. This icon is used to ilustrate our key results',
+  },
+  Scan: {
+    defaultMessage: 'Um ícone de scanner',
+    id: 'm8Extk',
+    description:
+      'The desc message for the Scan icon. This icon is used to ilustrate our key results',
+  },
+  Search: {
+    defaultMessage: 'Um ícone de busca',
+    id: 'c8UE59',
+    description:
+      'The desc message for the Search icon. This icon is used to ilustrate our key results',
+  },
+  TicketStar: {
+    defaultMessage: 'Um ícone de ingresso de cinema',
+    id: 'ZVTdlV',
+    description:
+      'The desc message for the TicketStar icon. This icon is used to ilustrate our key results',
+  },
+  TimesSquare: {
+    defaultMessage: 'Um ícone de relógio',
+    id: 'nJ7CuJ',
+    description:
+      'The desc message for the TimesSquare icon. This icon is used to ilustrate our key results',
+  },
+  Video: {
+    defaultMessage: 'Um ícone de video',
+    id: 'GHu/pe',
+    description:
+      'The desc message for the Video icon. This icon is used to ilustrate our key results',
+  },
+  Voice: {
+    defaultMessage: 'Um ícone de microfone',
+    id: '0xI/K7',
+    description:
+      'The desc message for the Voice icon. This icon is used to ilustrate our key results',
+  },
+  Wallet: {
+    defaultMessage: 'Um ícone de carteira',
+    id: 'lApj0K',
+    description:
+      'The desc message for the Wallet icon. This icon is used to ilustrate our key results',
+  },
+})
